03_train_and_predict/requirements.txt
jupyterlab==4.0.11
numpy==1.26.4
pandas==2.2.1
pyarrow==14.0.2
pyyaml==6.0.1
scikit-learn==1.4.2
seaborn==0.12.2